+++ crates/rust-client/src/transaction/mod.rs
@@ -172,6 +172,7 @@ pub use miden_protocol::transaction::{
 };
 pub use miden_protocol::vm::{AdviceInputs, AdviceMap};
 pub use miden_standards::account::interface::{AccountComponentInterface, AccountInterface};
+pub use miden_standards::tx_script::ExpirationTransactionScript;
 pub use miden_tx::auth::TransactionAuthenticator;
 pub use miden_tx::{
     DataStoreError,